change function location
