+++ maqianli/2024_training/collector_cpu.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
-	"os"
 	"strconv"
 	"strings"
 )
@@ -43,24 +41,6 @@ func collectTotalCPUTime(data []int) (int, error) {
 	return total, nil
 }
 
-func readFirstLineFromFile(filePath string) (string, error) {
-	file, err := os.Open(filePath)
-	if err != nil {
-		return "", fmt.Errorf("error opening file: %v", err)
-	}
-	defer file.Close()
-	scanner := bufio.NewScanner(file)
-	if scanner.Scan() {
-		return scanner.Text(), nil
-	}
-
-	if err = scanner.Err(); err != nil {
-		return "", fmt.Errorf("error reading file: %v", err)
-	}
-
-	return "", nil
-}
-
 func parseCPUUsageText(text string) ([]int, error) {
 	arr := strings.Fields(text)
 	if len(arr) != 11 {
@@ -87,18 +67,4 @@ func calculateCPUIdlePercentage(idle int, total int) (float64, error) {
 
 func collectIdleCPUTime(data []int) (int, error) {
 	return data[3], nil
-}
-
-func formatFloat(num float64, decimalPlaces int) float64 {
-	formatString := fmt.Sprintf("%%.%df", decimalPlaces)
-	formattedNum := fmt.Sprintf(formatString, num)
-
-	var result float64
-	_, err := fmt.Sscan(formattedNum, &result)
-	if err != nil {
-		fmt.Println("Error parsing formatted number:", err)
-		return 0.0
-	}
-
-	return result
 }
