+++ maqianli/2024_training/colletor_cpu_test.go
@@ -21,13 +21,6 @@ func TestCollectCPUUsageFromFile(t *testing.T) {
 	assert.Equal(t, 0.004553, data)
 }
 
-func TestReadFirstLineFromFile(t *testing.T) {
-	file := "linux_cpu_example.txt"
-	data, err := readFirstLineFromFile(file)
-	assert.Nil(t, err)
-	assert.Equal(t, "cpu  2310622 512 2961553 1154598147 4243 0 3471 0 0 0", data)
-}
-
 func TestParseCPUUsageText(t *testing.T) {
 	text := "cpu  2310622 512 2961553 1154598147 4243 0 3471 0 0 0"
 	data, err := parseCPUUsageText(text)
@@ -57,15 +50,4 @@ func TestCalculateCPUIdlePercentage(t *testing.T) {
 	got, err := calculateCPUIdlePercentage(idle, total)
 	assert.Nil(t, err)
 	assert.Equal(t, 0.995447, got)
-}
-
-func TestFormatFloat(t *testing.T) {
-	num := 0.002553
-	assert.Equal(t, 0.0, formatFloat(num, 0))
-	assert.Equal(t, 0.00, formatFloat(num, 2))
-	assert.Equal(t, 0.00, formatFloat(num, 2))
-	assert.Equal(t, 0.0026, formatFloat(num, 4))
-	assert.Equal(t, 0.0025530, formatFloat(num, 7))
-	assert.Equal(t, 0.00255300, formatFloat(num, 8))
-	assert.Equal(t, -0.00255300, formatFloat(-num, 8))
 }
